+++ cmd/deadeye/initrules.go
@@ -78,6 +78,7 @@ func runInitRules(host string, args []string) {
 		os.Exit(1)
 	}
 	fmt.Println(cGood("Wrote") + " " + path)
+	installPRCommand(host)
 	fmt.Println(cDim("Remove any time with: deadeye uninstall " + host))
 }
 
@@ -116,5 +117,6 @@ func runUninstallRules(host string) {
 		fmt.Fprintln(os.Stderr, "deadeye uninstall "+host+":", err)
 		os.Exit(1)
 	}
+	removePRCommand(host)
 	fmt.Println(cGood("Removed") + " " + path)
 }